fix : 병합 오류 수정
