+++ src/main/java/org/example/fourchak/domain/coupon/dto/CouponAdminResponseDto.java
@@ -1,0 +1,20 @@
+package org.example.fourchak.domain.coupon.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.example.fourchak.domain.coupon.entity.Coupon;
+
+@Getter
+@AllArgsConstructor
+public class CouponAdminResponseDto {
+
+    private Long couponId;
+    private int discount;
+    private int count;
+
+    public CouponAdminResponseDto(Coupon coupon) {
+        this.couponId = coupon.getId();
+        this.discount = coupon.getDiscount();
+        this.count = coupon.getCount();
+    }
+}
