+++ src/main/java/org/example/fourchak/domain/coupon/dto/CouponCreateRequestDto.java
@@ -1,0 +1,11 @@
+package org.example.fourchak.domain.coupon.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CouponCreateRequestDto {
+  private final int discount;
+  private final int count;
+}
