+++ src/main/java/org/example/fourchak/domain/coupon/dto/CouponUpdateRequestDto.java
@@ -1,0 +1,11 @@
+package org.example.fourchak.domain.coupon.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CouponUpdateRequestDto {
+
+    private int discount;
+}
