+++ src/main/java/org/example/fourchak/domain/reservation/service/ReservationService.java
@@ -1,0 +1,5 @@
+package org.example.fourchak.domain.reservation.service;
+
+public interface ReservationService {
+
+}
