+++ src/main/java/org/example/fourchak/domain/user/controller/UserController.java
@@ -1,0 +1,5 @@
+package org.example.fourchak.domain.user.controller;
+
+public class UserController {
+
+}
