+++ src/main/java/org/example/fourchak/domain/user/dto/response/UserResponse.java
@@ -1,0 +1,5 @@
+package org.example.fourchak.domain.user.dto.response;
+
+public class UserResponse {
+
+}
