+++ src/main/java/org/example/fourchak/domain/user/entity/User.java
@@ -1,0 +1,31 @@
+package org.example.fourchak.domain.user.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.example.fourchak.common.BaseEntity;
+
+@Getter
+@Entity
+@NoArgsConstructor
+@Table(name = "users")
+public class User extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String email;
+
+    private String username;
+
+    private String phone;
+
+    private String password;
+
+    private String role;
+}
