--- src/main/java/org/example/fourchak/reservation/dto/requset/ReservationRequestDto.java
@@ -1,20 +1,0 @@
-package org.example.fourchak.reservation.dto.requset;
-
-
-import java.time.LocalDateTime;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ReservationRequestDto {
-
-    private int peopleNumber;
-
-    private LocalDateTime reservationTime;
-
-    public ReservationRequestDto(int peopleNumber, LocalDateTime reservationTime) {
-        this.peopleNumber = peopleNumber;
-        this.reservationTime = reservationTime;
-    }
-}
