--- src/main/java/org/example/fourchak/store/repository/StoreRepository.java
@@ -1,13 +1,0 @@
-package org.example.fourchak.store.repository;
-
-import org.example.fourchak.store.entity.Store;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StoreRepository extends JpaRepository<Store, Long> {
-
-    default Store findStoreByIdOrElseThrow(Long id) {
-        return findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 가게입니다."));
-    }
-
-
-}
