--- src/main/java/org/example/fourchak/user/enums/UserRole.java
@@ -1,6 +1,0 @@
-package org.example.fourchak.user.enums;
-
-public enum UserRole {
-    USER, ADMIN;
-
-}
